backend/requirements.txt
flask
flask-cors
pillow
gunicorn
python-dotenv
google-generativeai
